requirements.txt
django==5.0.2
openai==1.12.0
google-generativeai==0.3.2
anthropic==0.8.1
python-dotenv==1.0.1 